requirements/common.txt
#common requirements
Django==1.5
djangorestframework
django-rest-framework-docs
